+++ CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Unreleased
 
+## 3.1.0.beta1
+
 ### openapi_first/test
 
 #### Changed
-- Test now raises a `OpenapiFirst::Test::UnknownQueryParameterError` when it sees unknown query parameters
+- Test now raises `OpenapiFirst::Test::UnknownQueryParameterError` when it sees unknown query parameters. Note that `OpenapiFirst` ("core") still allows unknown query parameters.
 - Test does not track requests/responses unless the OAD was registered via Test.register (or OpenapiFirst.register)
 
 ## 3.0.1